Distribution in stats
Normal distribution and uniform distribution are two commonly used probability distributions in statistics. 
A normal distribution is a bell-shaped curve that describes a set of random variables with a mean and a standard deviation.
A uniform distribution represents a set of random variables that are equally likely to occur within a given range.